# frontend/openapi.json
{
  "components": {
    "schemas": {
      "ApprovalRecommendation": {
        "properties": {
          "decided_by": {
            "default": "agent",
            "enum": [
              "agent",
              "policy"
            ],
            "title": "Decided By",
            "type": "string"
          },
          "final_route": {
            "$ref": "#/components/schemas/DecisionRoute"
          },
          "proposed_route": {
            "$ref": "#/components/schemas/DecisionRoute"
          },
          "reason_codes": {
            "items": {
              "type": "string"
            },
            "title": "Reason Codes",
            "type": "array"
          },
          "reflection_count": {
            "default": 0,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Reflection Count",
            "type": "integer"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "proposed_route",
          "final_route",
          "summary"
        ],
        "title": "ApprovalRecommendation",
        "type": "object"
      },
      "Body_create_run_api_v1_runs_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_create_run_api_v1_runs_post",
        "type": "object"
      },
      "DecisionRoute": {
        "enum": [
          "approve",
          "review",
          "reject"
        ],
        "title": "DecisionRoute",
        "type": "string"
      },
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "FindingSeverity": {
        "enum": [
          "info",
          "warning",
          "blocking"
        ],
        "title": "FindingSeverity",
        "type": "string"
      },
      "HealthResponse": {
        "properties": {
          "service": {
            "title": "Service",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "service"
        ],
        "title": "HealthResponse",
        "type": "object"
      },
      "HumanReview": {
        "properties": {
          "decided_at": {
            "format": "date-time",
            "title": "Decided At",
            "type": "string"
          },
          "decision": {
            "enum": [
              "approve",
              "reject"
            ],
            "title": "Decision",
            "type": "string"
          },
          "reason": {
            "maxLength": 500,
            "minLength": 3,
            "title": "Reason",
            "type": "string"
          },
          "resume_pending": {
            "default": true,
            "title": "Resume Pending",
            "type": "boolean"
          }
        },
        "required": [
          "decision",
          "reason",
          "decided_at"
        ],
        "title": "HumanReview",
        "type": "object"
      },
      "InvoiceData": {
        "properties": {
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "due_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "extraction_confidence": {
            "default": "high",
            "enum": [
              "high",
              "medium",
              "low"
            ],
            "title": "Extraction Confidence",
            "type": "string"
          },
          "invoice_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invoice Date"
          },
          "invoice_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invoice Number"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/InvoiceItem"
            },
            "title": "Items",
            "type": "array"
          },
          "payment_terms": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Terms"
          },
          "revision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision"
          },
          "shipping": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Money"
              },
              {
                "type": "null"
              }
            ]
          },
          "subtotal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Money"
              },
              {
                "type": "null"
              }
            ]
          },
          "tax": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Money"
              },
              {
                "type": "null"
              }
            ]
          },
          "total": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Money"
              },
              {
                "type": "null"
              }
            ]
          },
          "vendor_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vendor Name"
          }
        },
        "title": "InvoiceData",
        "type": "object"
      },
      "InvoiceItem": {
        "properties": {
          "line_number": {
            "minimum": 1.0,
            "title": "Line Number",
            "type": "integer"
          },
          "line_total": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Money"
              },
              {
                "type": "null"
              }
            ]
          },
          "normalized_item_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Normalized Item Code"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "source_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Name"
          },
          "unit_price": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Money"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "line_number"
        ],
        "title": "InvoiceItem",
        "type": "object"
      },
      "Money": {
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          }
        },
        "required": [
          "amount",
          "currency"
        ],
        "title": "Money",
        "type": "object"
      },
      "PaymentResult": {
        "properties": {
          "amount": {
            "$ref": "#/components/schemas/Money"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "mock_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mock Reference"
          },
          "status": {
            "enum": [
              "pending",
              "succeeded",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "status",
          "amount",
          "created_at",
          "updated_at"
        ],
        "title": "PaymentResult",
        "type": "object"
      },
      "ReviewRequest": {
        "properties": {
          "decision": {
            "enum": [
              "approve",
              "reject"
            ],
            "title": "Decision",
            "type": "string"
          },
          "reason": {
            "maxLength": 500,
            "minLength": 3,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "decision",
          "reason"
        ],
        "title": "ReviewRequest",
        "type": "object"
      },
      "RunCreationResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "deduplicated": {
            "title": "Deduplicated",
            "type": "boolean"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "source_filename": {
            "title": "Source Filename",
            "type": "string"
          },
          "stage": {
            "$ref": "#/components/schemas/RunStage"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "source_filename",
          "status",
          "stage",
          "created_at",
          "updated_at",
          "deduplicated"
        ],
        "title": "RunCreationResponse",
        "type": "object"
      },
      "RunDetail": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkflowError"
              },
              {
                "type": "null"
              }
            ]
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/RunEvent"
            },
            "title": "Events",
            "type": "array"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/ValidationFinding"
            },
            "title": "Findings",
            "type": "array"
          },
          "invoice": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InvoiceData"
              },
              {
                "type": "null"
              }
            ]
          },
          "payment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaymentResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "recommendation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApprovalRecommendation"
              },
              {
                "type": "null"
              }
            ]
          },
          "review": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HumanReview"
              },
              {
                "type": "null"
              }
            ]
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "source_filename": {
            "title": "Source Filename",
            "type": "string"
          },
          "stage": {
            "$ref": "#/components/schemas/RunStage"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "source_filename",
          "status",
          "stage",
          "created_at",
          "updated_at"
        ],
        "title": "RunDetail",
        "type": "object"
      },
      "RunEvent": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "event_id": {
            "title": "Event Id",
            "type": "integer"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "stage": {
            "$ref": "#/components/schemas/RunStage"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          }
        },
        "required": [
          "event_id",
          "stage",
          "status",
          "code",
          "message",
          "created_at"
        ],
        "title": "RunEvent",
        "type": "object"
      },
      "RunListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RunSummary"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "RunListResponse",
        "type": "object"
      },
      "RunStage": {
        "enum": [
          "ingest",
          "extract",
          "validate",
          "recommend",
          "review",
          "pay",
          "finalize"
        ],
        "title": "RunStage",
        "type": "string"
      },
      "RunStatus": {
        "enum": [
          "queued",
          "running",
          "review_required",
          "completed",
          "rejected",
          "failed"
        ],
        "title": "RunStatus",
        "type": "string"
      },
      "RunSummary": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "source_filename": {
            "title": "Source Filename",
            "type": "string"
          },
          "stage": {
            "$ref": "#/components/schemas/RunStage"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "source_filename",
          "status",
          "stage",
          "created_at",
          "updated_at"
        ],
        "title": "RunSummary",
        "type": "object"
      },
      "ValidationFinding": {
        "properties": {
          "actual": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Actual"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "expected": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Expected"
          },
          "field_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Path"
          },
          "item_line_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Line Number"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "severity": {
            "$ref": "#/components/schemas/FindingSeverity"
          }
        },
        "required": [
          "code",
          "severity",
          "message"
        ],
        "title": "ValidationFinding",
        "type": "object"
      },
      "WorkflowError": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "WorkflowError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Local asynchronous workspace for invoice processing.",
    "title": "Invoice Processing API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/health": {
      "get": {
        "operationId": "health_api_v1_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health",
        "tags": [
          "health"
        ]
      }
    },
    "/api/v1/runs": {
      "get": {
        "operationId": "list_runs_api_v1_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "List Runs",
        "tags": [
          "runs"
        ]
      },
      "post": {
        "operationId": "create_run_api_v1_runs_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_run_api_v1_runs_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunCreationResponse"
                }
              }
            },
            "description": "Deduplicated run"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunCreationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Create Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/runs/{run_id}": {
      "get": {
        "operationId": "get_run_api_v1_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/runs/{run_id}/review": {
      "post": {
        "operationId": "review_run_api_v1_runs__run_id__review_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Review Run",
        "tags": [
          "runs"
        ]
      }
    }
  }
}